requirements.txt
Flask==2.1.0
Werkzeug==2.0.3
Flask-Login==0.5.0
mysql-connector-python==8.0.26
dotenv==0.0.5